undo some renaming to reduce changes
