+++ frontend/src/lib/components/ChildrenGallery.svelte
